# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * SL-ERP API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthTokenResponse {
  token?: string;
  user_id?: number;
  username?: string;
  email?: string;
}

export interface AuthStatus {
  authenticated?: boolean;
  /** @nullable */
  username?: string | null;
}

export type UserProfileGroupsItem = {
  id?: number;
  name?: string;
};

export interface UserProfile {
  id?: number;
  username?: string;
  email?: string;
  first_name?: string;
  last_name?: string;
  is_superuser?: boolean;
  groups?: UserProfileGroupsItem[];
}

export interface UserProfileInput {
  first_name?: string;
  last_name?: string;
  email?: string;
}

export type PlatformUserGroupsItem = {
  id?: number;
  name?: string;
};

export interface PlatformUser {
  id?: number;
  username?: string;
  email?: string;
  first_name?: string;
  last_name?: string;
  is_active?: boolean;
  is_superuser?: boolean;
  groups?: PlatformUserGroupsItem[];
  date_joined?: string;
  /** @nullable */
  last_login?: string | null;
}

export interface RoleAssignmentInput {
  group_ids?: number[];
}

export type RolePermissionsItem = {
  id?: number;
  codename?: string;
  name?: string;
};

export interface Role {
  id?: number;
  name?: string;
  permissions?: RolePermissionsItem[];
}

export interface NavItem {
  key?: string;
  title?: string;
  icon?: string;
  route_path?: string;
  badge_text?: string;
  sort_order?: number;
}

export interface NavSection {
  key?: string;
  title?: string;
  icon?: string;
  sort_order?: number;
  items?: NavItem[];
}

export interface WorkspaceNavigation {
  sections?: NavSection[];
}

export interface MenuSection {
  id?: number;
  key?: string;
  title?: string;
  icon?: string;
  sort_order?: number;
  is_active?: boolean;
}

export interface MenuSectionInput {
  key: string;
  title: string;
  icon?: string;
  sort_order?: number;
  is_active?: boolean;
}

export interface MenuItem {
  id?: number;
  section?: number;
  key?: string;
  title?: string;
  icon?: string;
  route_path?: string;
  sort_order?: number;
  is_active?: boolean;
}

export interface MenuItemInput {
  section: number;
  key: string;
  title: string;
  icon?: string;
  route_path?: string;
  sort_order?: number;
  is_active?: boolean;
}

export interface Tenant {
  id?: number;
  name?: string;
  code?: string;
  legal_name?: string;
  contact_email?: string;
  contact_phone?: string;
  status?: string;
  is_active?: boolean;
  default_currency?: string;
  timezone?: string;
  created_at?: string;
}

export interface TenantInput {
  name: string;
  legal_name?: string;
  contact_email?: string;
  contact_phone?: string;
  status?: string;
  default_currency?: string;
  timezone?: string;
}

export interface ModuleDefinition {
  id?: number;
  slug?: string;
  name?: string;
  category?: string;
  description?: string;
  is_core?: boolean;
  is_active?: boolean;
}

export interface SubscriptionPlan {
  id?: number;
  name?: string;
  code?: string;
  billing_period?: string;
  price?: number;
  currency?: string;
  is_active?: boolean;
}

export interface SubscriptionPlanInput {
  name: string;
  billing_period?: string;
  price?: number;
  currency?: string;
}

export interface TenantSubscription {
  id?: number;
  tenant?: number;
  tenant_name?: string;
  plan?: number;
  plan_name?: string;
  status?: string;
  starts_at?: string;
  /** @nullable */
  expires_at?: string | null;
}

export interface IntegrationEndpoint {
  id?: number;
  tenant?: number;
  name?: string;
  transport_type?: string;
  provider?: string;
  base_url?: string;
  is_active?: boolean;
  is_primary?: boolean;
}

export interface IntegrationEndpointInput {
  tenant?: number;
  name: string;
  transport_type: string;
  provider?: string;
  base_url?: string;
  is_active?: boolean;
  is_primary?: boolean;
}

export type WeighbridgeDashboardTotals = {
  transactions_today?: number;
  transactions_this_month?: number;
  net_weight_today?: number;
  pending_payments?: number;
};

export type WeighbridgeDashboardStatusBreakdownItem = {
  status?: string;
  count?: number;
};

export interface Transaction {
  id?: number;
  branch?: number;
  branch_name?: string;
  customer?: number;
  customer_name?: string;
  vehicle?: number;
  vehicle_plate?: string;
  operator?: string;
  /** @nullable */
  item?: number | null;
  item_name?: string;
  /** @nullable */
  gross_weight?: number | null;
  /** @nullable */
  tare_weight?: number | null;
  /** @nullable */
  net_weight?: number | null;
  status?: string;
  weight_type?: string;
  payment_status?: string;
  /** @nullable */
  charge?: number | null;
  destination?: string;
  invoiced?: boolean;
  approval_status?: string;
  created_at?: string;
  updated_at?: string;
}

export interface WeighbridgeDashboard {
  totals?: WeighbridgeDashboardTotals;
  status_breakdown?: WeighbridgeDashboardStatusBreakdownItem[];
  recent_transactions?: Transaction[];
}

export interface TransactionInput {
  branch: number;
  customer: number;
  vehicle: number;
  item?: number;
  gross_weight?: number;
  tare_weight?: number;
  weight_type?: string;
  destination?: string;
  operator?: string;
}

export interface PaginatedTransactions {
  count?: number;
  /** @nullable */
  next?: string | null;
  /** @nullable */
  previous?: string | null;
  results?: Transaction[];
}

export interface Branch {
  id?: number;
  name?: string;
  address?: string;
  email?: string;
  phone?: string;
}

export interface Customer {
  id?: number;
  name?: string;
  email?: string;
  phone?: string;
}

export interface CustomerInput {
  name: string;
  email?: string;
  phone?: string;
  branch?: number;
}

export interface Vehicle {
  id?: number;
  number_plate?: string;
  customer?: number;
  customer_name?: string;
  /** @nullable */
  vehicle_type?: number | null;
  vehicle_type_name?: string;
}

export interface VehicleInput {
  number_plate: string;
  customer: number;
  vehicle_type?: number;
}

export interface LiveWeightReading {
  /** @nullable */
  weight?: number | null;
  unit?: string;
  stable?: boolean;
  source?: string;
  timestamp?: string;
}

export interface Invoice {
  id?: number;
  invoice_number?: string;
  customer?: number;
  customer_name?: string;
  /** @nullable */
  transaction?: number | null;
  status?: string;
  total_amount?: number;
  currency?: string;
  /** @nullable */
  issued_at?: string | null;
  /** @nullable */
  due_date?: string | null;
  created_at?: string;
}

export interface PaginatedInvoices {
  count?: number;
  /** @nullable */
  next?: string | null;
  /** @nullable */
  previous?: string | null;
  results?: Invoice[];
}

export interface ReceivePaymentInput {
  amount: number;
  method: number;
  reference?: string;
}

export interface PaymentMethod {
  id?: number;
  name?: string;
  is_active?: boolean;
}

export type PaymentSummaryInvoicesByStatusItem = {
  status?: string;
  count?: number;
  total?: number;
};

export interface PaymentSummary {
  total_invoiced?: number;
  total_received?: number;
  outstanding?: number;
  invoices_by_status?: PaymentSummaryInvoicesByStatusItem[];
}

export type GetWeighbridgeDashboardParams = {
tenant_code?: string;
branch_id?: number;
};

export type ListTransactionsParams = {
tenant_code?: string;
branch_id?: number;
status?: string;
payment_status?: string;
page?: number;
};

export type ListCustomersParams = {
tenant_code?: string;
branch_id?: number;
search?: string;
};

export type ListVehiclesParams = {
customer_id?: number;
search?: string;
};

export type GetLiveWeightParams = {
branch_id?: number;
};

export type ListInvoicesParams = {
tenant_code?: string;
status?: string;
page?: number;
};

export type GetPaymentSummaryParams = {
tenant_code?: string;
};