# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SL-ERP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/invoices/{id}/receive-payment/:
    post:
      operationId: receivePayment
      tags: [payments]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivePaymentInput"
      responses:
        "200":
          description: Payment received
...
components:
  schemas:
...
    ReceivePaymentInput:
      type: object
      required: [amount, method]
      properties:
        amount: { type: number }
        method: { type: integer }
        reference: { type: string }
...